simulazione3/challs/WeirdCPU/src/requirements.txt
Flask>=2,<3
requests>=2,<3
python-dotenv>=0.18,<0.19
